Rename tables for sanity
Rename TxMetadata to PoolMetadata because that is the actual pool
metadata. It's nothing to do with transactions (and it's really
confusing since Shelley also really does have transaction metadata).

Rename PoolMetaData to PoolMetadataReference because this is _not_ the
pool metadata, but the reference/pointer to the pool metadata. It has
the URL and expected hash of the actual pool metadata.