Add asynchronous updating of offline metadata
All offline metadata is tried synchronously with block retrieval. Requests
that fail are then tried asynchonously. Requests than need to be retried
are queried from the database and retries are back off exponentially with
a ceiling of daily retries. The retry interval for all offline metadata
is reduced to 60 seconds when the program is restarted.